requirements.txt
Django==5.1.2
psycopg[binary]==3.3.1
django-environ==0.11.2
djangorestframework==3.15.2
celery==5.4.0
redis==5.0.1
python-telegram-bot==21.7
spotipy==2.23.0
requests==2.31.0
